Moved MainView to main package and removed javadoc author.
