+++ android/app/src/main/kotlin/com/chicagoroboto/features/TabHolder.kt
@@ -2,9 +2,6 @@ package com.chicagoroboto.features
 
 import android.support.design.widget.TabLayout
 
-/**
- * Created by feliperoriz on 4/7/18.
- */
 interface TabHolder {
     var tabLayout: TabLayout?
 }
