+++ android/app/src/main/kotlin/com/chicagoroboto/features/sessions/SessionListView.kt
@@ -5,9 +5,7 @@ import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
 import android.util.AttributeSet
 import android.view.ViewGroup
-import com.chicagoroboto.R
 import com.chicagoroboto.ext.getComponent
-import com.chicagoroboto.features.MainView
 import com.chicagoroboto.features.main.MainComponent
 import com.chicagoroboto.model.Session
 import com.chicagoroboto.model.Speaker
